refactor: extract anthropic handler in base AYC-63
